requirements.txt
asgiref==3.5.0
dj-static==0.0.6
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
gunicorn==20.1.0
numpy==1.22.2
opencv-python==4.5.5.62
Pillow==9.0.1
pytz==2021.3
qrcode==7.3.1
sqlparse==0.4.2
static3==0.7.0